Actualizado en varios lugares el instalador y el lugar donde se guardan los archivos q se suben

+++ docs/instalador/instalar.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-version=0.4
+version=0.9
 if [ $(uname -a|grep x86_64|wc -l) -gt 0 ];
   then 
      versionKernel=64;
@@ -114,6 +114,12 @@ generarJaula()
   echo "Generando el certificado de apache"
   mkdir -p /etc/apache2/ssl/$ID
   openssl req -x509 -nodes -days 3650 -newkey rsa:2048 -keyout /etc/apache2/ssl/$ID/apache.pem -out /etc/apache2/ssl/$ID/apache.pem
+  a2enmod rewrite
+  a2enmod expires
+  a2enmod ssl
+  a2enmod headers
+  echo "Procederemos a habilitar en apache los sites"
+  a2dissite default
   a2ensite $ID-apache-jaula-ssl
   a2ensite $ID-apache-jaula-opac
 }
@@ -132,8 +138,13 @@ actualizarBDD()
 		echo 'Luego debe aplicar el sql que esta en ~/pendiente/UPDATES_MERAN_SANDBOX.sql para actualizar su base de datos'
 	else
 		echo 'Base de datos actualizada con exito'
-	fi
-	rm /tmp/UPDATES_MERAN_SANDBOX.sql
+   fi
+   if [ $CONSERVE_SQL -eq 0 ]
+	then
+		rm /tmp/UPDATES_MERAN_SANDBOX.sql
+	else
+		mv /tmp/UPDATES_MERAN_SANDBOX.sql ~/pendiente/
+   fi
 }
 
 
@@ -159,7 +170,11 @@ generarPermisosBDD()
 	else
 		echo 'Base de datos creada con exito'
 	fi
-  rm /tmp/$ID.permisosbdd*
+ if [ $CONSERVE_SQL -eq 1 ]
+	then
+		mv /tmp/$ID.permisosbdd ~/pendiente/base.sql
+	fi
+ rm /tmp/$ID.permisosbdd*
 }
 
 generarConfiguracion()
@@ -208,12 +223,6 @@ instalarDependencias()
         apt-get install apache2 apache2-mpm-prefork mysql-server libapache2-mod-perl2 libgd2-xpm libxpm4 htmldoc libaspell15 ntpdate -y
         #Configurar apache
         echo "Procederemos a habilitar en apache los modulos necesarios"
-        a2enmod rewrite
-        a2enmod expires
-        a2enmod ssl
-        a2enmod headers
-        echo "Procederemos a habilitar en apache los sites"
-        a2dissite default
 	}
 generarScriptDeInicio()
 {  
@@ -277,7 +286,9 @@ OPTIONS:
    -P	   La pass del usuario para crear la base de datos
    -U	   El usuario para conectarse a la base de datos
    -N	   Si esta presente Determina si es una instalacion nueva. Por defecto pregunta
-   -B	   Si esta presente Determina que tiene que instalr las dependencias. Por defecto pregunta
+   -B	   Si esta presente en 1 determina que tiene que instalar las dependencias, si esta en 0 no instala y si no esta pregunta.
+   -v	   Si esta presente determina la version del kernel que se va a utilizar. Por defecto la descubre de uname -a
+   -C	   Si esta presente indica que hay que consevar el sql de la base de datos sin eliminar
 EOF
 }
 
@@ -301,22 +312,27 @@ ROOT_USER_BASE="root"
 ROOT_PASS_BASE=""
 DBB_USER_ORIGEN="localhost"
 NEW_INSTALACION=0
-BASE_INSTALACION=0
+BASE_INSTALACION=2
+CONSERVE_SQL=0
 
-while getopts “?:h:i:d:b:u:p:s:w:c:U:P:NB” OPTION
+while getopts “?:h:i:d:b:u:p:s:w:c:U:v:P:NB:C” OPTION
 do
      case $OPTION in
          U)
 			ROOT_USER_BASE=$OPTARG
 			;;
-		 N)
+	 N)
 			NEW_INSTALACION=1
 			;;
-		 B)
-			BASE_INSTALACION=1
+	 B)
+			BASE_INSTALACION=$OPTARG
 			;;
-		 P)	
+	 P)	
 			ROOT_PASS_BASE=$OPTARG
+            ;;
+	 v)	
+
+			versionKernel=$OPTARG
             ;;
          i)
              ID=$OPTARG
@@ -345,6 +361,9 @@ do
              ;;
          c)
              CONFIGURACION_MERAN=$OPTARG
+             ;;
+         C)
+             CONSERVE_SQL=1
              ;;
          ?)
              usage
@@ -384,9 +403,11 @@ if [ $BASE_INSTALACION -eq 1 ]
 then
 	instalarDependencias
 else
-	echo "¿Quiere proceder a instalar todo el software de base base necesaria para Meran? (Apache/Mysql/perl/etc)"
-	select OPCION in Instalar No_instalar 
-	do
+	if [ $BASE_INSTALACION -eq 2 ] 
+	then
+		echo "¿Quiere proceder a instalar todo el software de base base necesaria para Meran? (Apache/Mysql/perl/etc)"
+		select OPCION in Instalar No_instalar 
+		do
 		if [ $OPCION = Instalar ]; 
 			then
 			instalarDependencias
@@ -395,7 +416,8 @@ else
 			echo "NO se instalará nada base"
 				break
 		fi
-	done
+		done
+	fi	
 fi
 
 echo "Ahora si vamos a instalar Meran en el sistema"
